refactor(ChallengesForm): adding real data, need to refactor overall somewhat as one cannot currently create your own challenge, but rather needs to make use of the challenges provided
