+++ src/modules/Challenges/Challenges.selectors.ts
@@ -1,0 +1,3 @@
+import { RootState } from '../../redux/redux.types'
+
+export const selectChallenges = (state: RootState) => state.challenges
