+++ src/modules/CompletedChallenges/Form/CompletedChallengesForm.styles.ts
