+++ src/modules/CompletedChallenges/Form/CompletedChallengesFrom.selector.ts
@@ -1,0 +1,5 @@
+import { createSelector } from '@reduxjs/toolkit'
+
+import { selectChallenges } from '../../Challenges/Challenges.selectors'
+
+export default createSelector(selectChallenges, challenges => ({ challenges }))
